refactor: move business logic to a custom hook
